# src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UpdateUserOutput": {
				"properties": {
					"user": {
						"properties": {
							"updated_at": {
								"type": "string"
							},
							"created_at": {
								"type": "string"
							},
							"userProfile": {
								"properties": {
									"followersCount": {
										"type": "number",
										"format": "double"
									},
									"followingCount": {
										"type": "number",
										"format": "double"
									},
									"links": {
										"items": {
											"properties": {
												"url": {
													"type": "string"
												},
												"id": {
													"type": "string"
												}
											},
											"required": [
												"url",
												"id"
											],
											"type": "object"
										},
										"type": "array"
									},
									"backgroundImage": {
										"type": "string",
										"nullable": true
									},
									"icon": {
										"type": "string",
										"nullable": true
									},
									"autobiography": {
										"type": "string",
										"nullable": true
									},
									"displayName": {
										"type": "string",
										"nullable": true
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"followersCount",
									"followingCount",
									"links",
									"backgroundImage",
									"icon",
									"autobiography",
									"displayName",
									"id"
								],
								"type": "object",
								"nullable": true
							},
							"verified": {
								"type": "boolean"
							},
							"email": {
								"type": "string"
							},
							"username": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"updated_at",
							"created_at",
							"userProfile",
							"verified",
							"email",
							"username",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Express.Multer.File": {
				"description": "Object containing file metadata and access information.",
				"properties": {
					"fieldname": {
						"type": "string",
						"description": "Name of the form field associated with this file."
					},
					"originalname": {
						"type": "string",
						"description": "Name of the file on the uploader's computer."
					},
					"encoding": {
						"type": "string",
						"description": "Value of the `Content-Transfer-Encoding` header for this file.",
						"deprecated": true
					},
					"mimetype": {
						"type": "string",
						"description": "Value of the `Content-Type` header for this file."
					},
					"size": {
						"type": "number",
						"format": "double",
						"description": "Size of the file in bytes."
					},
					"stream": {
						"type": "string",
						"format": "byte",
						"description": "A readable stream of this file. Only available to the `_handleFile`\ncallback for custom `StorageEngine`s."
					},
					"destination": {
						"type": "string",
						"description": "`DiskStorage` only: Directory to which this file has been uploaded."
					},
					"filename": {
						"type": "string",
						"description": "`DiskStorage` only: Name of this file within `destination`."
					},
					"path": {
						"type": "string",
						"description": "`DiskStorage` only: Full path to the uploaded file."
					},
					"buffer": {
						"type": "string",
						"format": "byte",
						"description": "`MemoryStorage` only: A Buffer containing the entire file."
					}
				},
				"required": [
					"fieldname",
					"originalname",
					"encoding",
					"mimetype",
					"size",
					"stream",
					"destination",
					"filename",
					"path",
					"buffer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserInput": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"autobiography": {
						"type": "string"
					},
					"links": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"icon": {
						"$ref": "#/components/schemas/Express.Multer.File"
					},
					"backgroundImage": {
						"$ref": "#/components/schemas/Express.Multer.File"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
...
		"/v1/users": {
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "User updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UpdateUserOutput"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"description": "Update authenticated user's profile information\n\nThis endpoint allows authenticated users to update their profile information\nincluding basic details and profile images. It supports partial updates,\nfile uploads for profile images, and comprehensive validation.\n\nUpdate Features:\n- Update basic profile information (displayName, autobiography)\n- Upload and manage profile icon (with size and type validation)\n- Upload and manage background image\n- Automatic file cleanup for replaced images\n- Secure file storage via S3Gateway\n- Comprehensive input validation",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- User update data including profile fields and optional files",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserInput",
								"description": "- User update data including profile fields and optional files"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Users retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pagination": {
											"properties": {
												"total": {
													"type": "number",
													"format": "double"
												},
												"limit": {
													"type": "number",
													"format": "double"
												},
												"page": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"total",
												"limit",
												"page"
											],
											"type": "object"
										},
										"users": {
											"items": {},
											"type": "array"
										}
									},
									"required": [
										"pagination",
										"users"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieve paginated list of users with search functionality\n\nThis endpoint provides a paginated list of all users in the system with\noptional search functionality. It's useful for user discovery, admin\nmanagement, and general user browsing features.\n\nFeatures:\n- Pagination with configurable page size\n- Search functionality across user fields\n- Serialized user data with profile information\n- Comprehensive pagination metadata\n- Performance optimized queries",
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"description": "- Page number for pagination (default: 0)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Number of users per page (default: 10)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Optional search term to filter users",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "User deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Permanently delete authenticated user's account and associated data\n\nThis endpoint allows authenticated users to permanently delete their account\nand all associated data. It performs comprehensive cleanup including user\nfiles, database records, and external service data. This action is irreversible.\n\nDeletion Process:\n- Validates user authentication and ownership\n- Retrieves user file information (profile images, etc.)\n- Deletes user files from cloud storage (S3)\n- Removes all user data from database\n- Cleans up related records and associations\n- Comprehensive error handling and rollback\n\nSecurity Features:\n- JWT token validation required\n- User can only delete their own account\n- Comprehensive data cleanup\n- Secure file deletion from cloud storage",
				"tags": [
					"User Management"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/v1/users/{userId}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "User retrieved successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Retrieve specific user by ID with complete profile information\n\nThis endpoint retrieves detailed information about a specific user by their ID.\nIt returns complete user data including profile information, links, and\nfollower/following counts. Useful for user profiles, user details pages,\nand user lookup functionality.\n\nFeatures:\n- Complete user profile information\n- Social links and connections data\n- Follower and following counts\n- Serialized user data format\n- Proper error handling for non-existent users",
				"tags": [
					"User Management"
				],
				"security": [],
				"parameters": [
					{
						"description": "- Unique identifier of the user to retrieve",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}